fix: up-to-date working correctly
There was a problem when I rebuild module separatelly and then, when I want to rebuild it via some dependency, the module was marked as not-build and artisan want to build it again.

The correct behavior should be: If module was built once and no change was made, the build must be 'up-to-date'.

The problem was in incorrectly computed hash for Ref. The problem was the Ref wasn't trimmed and '//my/module:build   ' have different hash as '//my/module:build'.